ns/trial: removed client rx

+++ crates/tests/src/notification_service.rs
@@ -45,7 +45,10 @@ async fn generate_and_add_notifications() -> anyhow::Result<()> {
     app_state
         .redis_pool
         .xadd(
-            &notification_client_key(CLIENT_ID, &(hash_uuid(CLIENT_ID) % app_state.max_shards)),
+            &notification_client_key(
+                CLIENT_ID,
+                &((hash_uuid(CLIENT_ID) % app_state.max_shards as u128) as u64),
+            ),
             data.to_vec(),
             1000,
         )
@@ -56,7 +59,7 @@ async fn generate_and_add_notifications() -> anyhow::Result<()> {
         .xread(
             [notification_client_key(
                 CLIENT_ID,
-                &(hash_uuid(CLIENT_ID) % app_state.max_shards),
+                &((hash_uuid(CLIENT_ID) % app_state.max_shards as u128) as u64),
             )]
             .to_vec(),
             ["0-0".to_string()].to_vec(),
